Fix the use of USB gamepads.
Regrettably, this does require we pass in all devices to the flatpak environment, but it seems to be pretty standard practice among all games and game tools at the moment.

As a driveby, capture the runtime-version as a string, fixing a warning in the flatpak-builder about certain yaml parsers interpreting it as a number.